Codes/01/requirements.txt
tensorflow>=2.12,<2.16
numpy>=1.24,<2.0
pandas>=2.0
matplotlib>=3.7
scipy>=1.10
scikit-learn>=1.2
scikit-image>=0.21
pyyaml>=6.0
tqdm>=4.65
opencv-python-headless>=4.7
tensorboard>=2.12
Pillow>=9.5
